Fixed enum to int cast

+++ Robotix/External/PwmPin.cs
@@ -37,7 +37,7 @@ namespace Robotix.External
 		{
 			PwmValue = value;
 			PwmRange = range;
-			WiringPiSharp.SoftPwm.SoftPwmCreate (pin, value, range);
+			WiringPiSharp.SoftPwm.SoftPwmCreate ((int)pin, value, range);
 		}
 
 		/// <summary>
@@ -80,13 +80,13 @@ namespace Robotix.External
 		{
 			if (value == true)
 			{
-				SoftPwm.SoftPwmWrite (PhysicalPin, PwmValue);
+				SoftPwm.SoftPwmWrite ((int)PhysicalPin, PwmValue);
 				CurrentState = value;
 				JustChanged = true;
 			}
 			else
 			{
-				SoftPwm.SoftPwmWrite (PhysicalPin, 0);
+				SoftPwm.SoftPwmWrite ((int)PhysicalPin, 0);
 				CurrentState = value;
 				JustChanged = true;
 			}
@@ -97,7 +97,7 @@ namespace Robotix.External
 		/// <param name="pwmValue">Value to write</param>
 		public virtual void Write(int pwmValue)
 		{
-			SoftPwm.SoftPwmWrite (PhysicalPin, pwmValue);
+			SoftPwm.SoftPwmWrite ((int)PhysicalPin, pwmValue);
 			PwmValue = pwmValue;
 			CurrentState = true;
 			JustChanged = true;
